Swap to switch statement for readability, rework operator button logic
Previous code did not allow for chaining of operators. Now, chaining
an operator calculates the current operation and then stores it as the
first operand when pressing another operator button. The operate
function displays the result but now also nulls the displayValue
afterwards. This prevents the previous unintended behavior where
pressing an operator button would then use the displayed result as the
second operand as well as the first. Pressing number buttons now also
clears the previous result instead of concatenatig to it.