Implement a 'clear' flag to fix a bug
After calculating an expression with the equals button, the result is
automatically set as the first operand for further calculations. This
is intended behavior when the next input is an operator button. But, if
it's a number button, the intention should be to start a new
calculation. The flag allows wiping of the operands only if a number
button is pressed directly after a calculation is finished.