Add a secondary display and a clearPending flag
The flag is set to true after any time the operate function is called.
This allows for different actions based on the button pressed next, such
as clearing the display if it's a number button, but using the result
displayed from operate as the first operand if an operator button is
pressed instead.